+++ modules/apps/collaboration/message-boards/message-boards-web/src/main/resources/META-INF/resources/message_boards/init.jsp
@@ -84,8 +84,8 @@ page import="com.liferay.message.boards.kernel.service.MBThreadFlagLocalServiceU
 page import="com.liferay.message.boards.kernel.service.MBThreadLocalServiceUtil" %><%@
 page import="com.liferay.message.boards.kernel.service.MBThreadServiceUtil" %><%@
 page import="com.liferay.message.boards.kernel.util.comparator.MessageCreateDateComparator" %><%@
-page import="com.liferay.message.boards.util.comparator.MBObjectModifiedDateComparator" %><%@
-page import="com.liferay.message.boards.util.comparator.MBObjectTitleComparator" %><%@
+page import="com.liferay.message.boards.util.comparator.CategoryTitleComparator" %><%@
+page import="com.liferay.message.boards.util.comparator.ThreadModifiedDateComparator" %><%@
 page import="com.liferay.message.boards.web.constants.MBPortletKeys" %><%@
 page import="com.liferay.message.boards.web.internal.dao.search.MBResultRowSplitter" %><%@
 page import="com.liferay.message.boards.web.internal.display.context.MBDisplayContextProvider" %><%@
