+++ modules/apps/collaboration/message-boards/message-boards-web/src/main/resources/META-INF/resources/message_boards/view.jsp
@@ -70,7 +70,7 @@ if (orderByType.equals("asc")) {
 OrderByComparator orderByComparator = null;
 
 if (orderByCol.equals("modified-date")) {
-	orderByComparator = new MBObjectModifiedDateComparator(orderByAsc);
+	orderByComparator = new ThreadModifiedDateComparator(orderByAsc);
 
 }
